+++ test/internals/inputs/test-functional_response.jl
@@ -80,40 +80,40 @@ end
 @testset "Bioenergetic functional response functor" begin
     # Index by index - FoodWeb
     Fbioenergetic1_fw = BioenergeticResponse(foodweb_2sp)
-    @test Fbioenergetic1_fw([1, 1], 1, 1) == 0 # no interaction
-    @test Fbioenergetic1_fw([1, 1], 1, 2) == 0 # no interaction
-    @test Fbioenergetic1_fw([1, 1], 2, 2) == 0 # no interaction
-    @test Fbioenergetic1_fw([1, 1], 2, 1) == 1^2 / (0.5^2 + 1^2) # interaction
-    @test Fbioenergetic1_fw([1, 2], 2, 1) == 1^2 / (0.5^2 + 1^2) # don't depend on cons. B
-    @test Fbioenergetic1_fw([2, 1], 2, 1) == 2^2 / (0.5^2 + 2^2) # ...but depend on res. B
+    @test Fbioenergetic1_fw([1.0, 1], 1, 1) == 0 # no interaction
+    @test Fbioenergetic1_fw([1.0, 1], 1, 2) == 0 # no interaction
+    @test Fbioenergetic1_fw([1.0, 1], 2, 2) == 0 # no interaction
+    @test Fbioenergetic1_fw([1.0, 1], 2, 1) == 1^2 / (0.5^2 + 1^2) # interaction
+    @test Fbioenergetic1_fw([1.0, 2], 2, 1) == 1^2 / (0.5^2 + 1^2) # don't depend on cons. B
+    @test Fbioenergetic1_fw([2.0, 1], 2, 1) == 2^2 / (0.5^2 + 2^2) # ...but depend on res. B
 
     # Index by index - FoodWeb
     Fbioenergetic1_nti = BioenergeticResponse(multi_net1)
-    @test Fbioenergetic1_nti([1, 1], 1, 1) == 0 # no interaction
-    @test Fbioenergetic1_nti([1, 1], 1, 2) == 0 # no interaction
-    @test Fbioenergetic1_nti([1, 1], 2, 2) == 0 # no interaction
-    @test Fbioenergetic1_nti([1, 1], 2, 1) == 1^2 / (0.5^2 + 1^2) # interaction
-    @test Fbioenergetic1_nti([1, 2], 2, 1) == 1^2 / (0.5^2 + 1^2) # don't depend on cons. B
-    @test Fbioenergetic1_nti([2, 1], 2, 1) == 2^2 / (0.5^2 + 2^2) # ...but depend on res. B
+    @test Fbioenergetic1_nti([1.0, 1], 1, 1) == 0 # no interaction
+    @test Fbioenergetic1_nti([1.0, 1], 1, 2) == 0 # no interaction
+    @test Fbioenergetic1_nti([1.0, 1], 2, 2) == 0 # no interaction
+    @test Fbioenergetic1_nti([1.0, 1], 2, 1) == 1^2 / (0.5^2 + 1^2) # interaction
+    @test Fbioenergetic1_nti([1.0, 2], 2, 1) == 1^2 / (0.5^2 + 1^2) # don't depend on cons. B
+    @test Fbioenergetic1_nti([2.0, 1], 2, 1) == 2^2 / (0.5^2 + 2^2) # ...but depend on res. B
 
     # Matrix
     F21 = 1^2 / (0.5^2 + 1^2)
-    @test Fbioenergetic1_fw([1, 1]) == sparse([0 0; F21 0]) # provide biomass vector
-    @test Fbioenergetic1_fw([1, 1], foodweb_2sp) == sparse([0 0; F21 0]) # meth. consistency
-    @test Fbioenergetic1_fw([1, 1], multi_net1) == sparse([0 0; F21 0]) # meth. consistency
-    @test Fbioenergetic1_fw(1) == sparse([0 0; F21 0]) # or a scalar if same for all sp
+    @test Fbioenergetic1_fw([1.0, 1]) == sparse([0 0; F21 0]) # provide biomass vector
+    @test Fbioenergetic1_fw([1.0, 1], foodweb_2sp) == sparse([0 0; F21 0]) # meth. consistency
+    @test Fbioenergetic1_fw([1.0, 1], multi_net1) == sparse([0 0; F21 0]) # meth. consistency
+    @test Fbioenergetic1_fw(1.0) == sparse([0 0; F21 0]) # or a scalar if same for all sp
 
     # Non-default hill exponent
     Fbioenergetic_1 = BioenergeticResponse(foodweb_2sp; h = 3)
     F21 = 2^3 / (0.5^3 + 2^3)
-    @test Fbioenergetic_1(2) == sparse([0 0; F21 0])
-    @test Fbioenergetic_1(2, foodweb_2sp) == sparse([0 0; F21 0]) # method consistency
-    @test Fbioenergetic_1(2, multi_net1) == sparse([0 0; F21 0]) # nti=0 ⟺ foodweb
+    @test Fbioenergetic_1(2.0) == sparse([0 0; F21 0])
+    @test Fbioenergetic_1(2.0, foodweb_2sp) == sparse([0 0; F21 0]) # method consistency
+    @test Fbioenergetic_1(2.0, multi_net1) == sparse([0 0; F21 0]) # nti=0 ⟺ foodweb
 
     # Consumer feeding on several resources
     Fbioenergetic2_fw = BioenergeticResponse(foodweb_3sp)
     Fbioenergetic2_nti = BioenergeticResponse(net_interference)
-    B = [1, 1, 1] # uniform biomass distribution
+    B = [1.0, 1, 1] # uniform biomass distribution
     F21 = 1^2 / (0.5^2 + 1^2)
     F31 = 0.5 * 1^2 / ((0.5^2) + (0.5 * 1^2) + (0.5 * 1^2))
     F32 = F31
@@ -121,7 +121,7 @@ end
     @test Fbioenergetic2_nti(B) == sparse([0 0 0; F21 0 0; F31 F32 0])
     @test Fbioenergetic2_fw(B, foodweb_3sp) == sparse([0 0 0; F21 0 0; F31 F32 0])
     @test Fbioenergetic2_nti(B, net_interference) == sparse([0 0 0; F21 0 0; F31 F32 0])
-    B = [3, 2, 1] # non-uniform biomass distribution
+    B = [3.0, 2, 1] # non-uniform biomass distribution
     F21 = 3^2 / (0.5^2 + 3^2)
     F31 = 0.5 * 3^2 / ((0.5^2) + (0.5 * 3^2) + (0.5 * 2^2))
     F32 = 0.5 * 2^2 / ((0.5^2) + (0.5 * 3^2) + (0.5 * 2^2))
@@ -133,7 +133,7 @@ end
     # Adding intraspecific interference
     Fbioenergetic2_fw = BioenergeticResponse(foodweb_3sp; c = 1)
     Fbioenergetic2_nti = BioenergeticResponse(net_interference; c = 1)
-    B = [3, 2, 1] # non-uniform biomass distribution
+    B = [3.0, 2, 1] # non-uniform biomass distribution
     F21 = 3^2 / (0.5^2 + 1 * 2 * 0.5^2 + 3^2)
     F31 = 0.5 * 3^2 / ((0.5^2) + (1 * 1 * 0.5^2) + (0.5 * 3^2 + 0.5 * 2^2))
     F32 = 0.5 * 2^2 / ((0.5^2) + (1 * 1 * 0.5^2) + (0.5 * 3^2 + 0.5 * 2^2))
@@ -146,24 +146,24 @@ end
 @testset "Linear functional response functor" begin
     # Index by index - FoodWeb
     Flinear1_fw = LinearResponse(foodweb_2sp)
-    @test Flinear1_fw([1, 1], 1, 1) == 0 # no interaction
-    @test Flinear1_fw([1, 1], 1, 2) == 0 # no interaction
-    @test Flinear1_fw([1, 1], 2, 2) == 0 # no interaction
-    @test Flinear1_fw([1, 1], 2, 1) == 1.0 # interaction
-    @test Flinear1_fw([1, 2], 2, 1) == 1.0 # don't depend on cons. mass
-    @test Flinear1_fw([2, 1], 2, 1) == 2.0 # ...but depend on res. mass
+    @test Flinear1_fw([1.0, 1], 1, 1) == 0 # no interaction
+    @test Flinear1_fw([1.0, 1], 1, 2) == 0 # no interaction
+    @test Flinear1_fw([1.0, 1], 2, 2) == 0 # no interaction
+    @test Flinear1_fw([1.0, 1], 2, 1) == 1.0 # interaction
+    @test Flinear1_fw([1.0, 2], 2, 1) == 1.0 # don't depend on cons. mass
+    @test Flinear1_fw([2.0, 1], 2, 1) == 2.0 # ...but depend on res. mass
 
     # Index by index - MultiplexNetwork
     Flinear1_nti = LinearResponse(multi_net1)
-    @test Flinear1_nti([1, 1], 1, 1) == 0 # no interaction
-    @test Flinear1_nti([1, 1], 1, 2) == 0 # no interaction
-    @test Flinear1_nti([1, 1], 2, 2) == 0 # no interaction
-    @test Flinear1_nti([1, 1], 2, 1) == 1.0 # interaction
-    @test Flinear1_nti([1, 2], 2, 1) == 1.0 # don't depend on cons. mass
-    @test Flinear1_nti([2, 1], 2, 1) == 2.0 # ...but depend on res. mass
+    @test Flinear1_nti([1.0, 1], 1, 1) == 0 # no interaction
+    @test Flinear1_nti([1.0, 1], 1, 2) == 0 # no interaction
+    @test Flinear1_nti([1.0, 1], 2, 2) == 0 # no interaction
+    @test Flinear1_nti([1.0, 1], 2, 1) == 1.0 # interaction
+    @test Flinear1_nti([1.0, 2], 2, 1) == 1.0 # don't depend on cons. mass
+    @test Flinear1_nti([2.0, 1], 2, 1) == 2.0 # ...but depend on res. mass
 
     # Matrix
-    @test Flinear1_fw([1, 1]) == sparse([0 0; 1.0 0]) # provide biomass vector
+    @test Flinear1_fw([1.0, 1]) == sparse([0 0; 1.0 0]) # provide biomass vector
     @test Flinear1_fw(1) == sparse([0 0; 1.0 0]) # or a scalar if same for all sp
     @test Flinear1_nti([1, 1]) == sparse([0 0; 1.0 0]) # provide biomass vector
     @test Flinear1_nti(1) == sparse([0 0; 1.0 0]) # or a scalar if same for all sp
@@ -187,74 +187,74 @@ end
 @testset "Classic functional response functor" begin
     # Index by index - FoodWeb
     Fclassic1_fw = ClassicResponse(foodweb_2sp; hₜ = 1.0, aᵣ = 0.5)
-    @test Fclassic1_fw([1, 1], 1, 1, 1) == 0 # no interaction
-    @test Fclassic1_fw([1, 1], 1, 2, 1) == 0 # no interaction
-    @test Fclassic1_fw([1, 1], 2, 2, 1) == 0 # no interaction
+    @test Fclassic1_fw([1.0, 1], 1, 1, 1) == 0 # no interaction
+    @test Fclassic1_fw([1.0, 1], 1, 2, 1) == 0 # no interaction
+    @test Fclassic1_fw([1.0, 1], 2, 2, 1) == 0 # no interaction
     F21 = (1 * 0.5 * 1^2) / (1 + 0.5 * 1 * 1^2)
-    @test Fclassic1_fw([1, 1], 2, 1, 1) == F21 # interaction
-    @test Fclassic1_fw([1, 2], 2, 1, 1) == F21 # don't depend on cons. mass
-    @test Fclassic1_fw([1, 1], 2, 1, 2) == F21 / 2
-    @test Fclassic1_fw([1, 1], 2, 1, 3) == F21 / 3
+    @test Fclassic1_fw([1.0, 1], 2, 1, 1) == F21 # interaction
+    @test Fclassic1_fw([1.0, 2], 2, 1, 1) == F21 # don't depend on cons. mass
+    @test Fclassic1_fw([1.0, 1], 2, 1, 2) == F21 / 2
+    @test Fclassic1_fw([1.0, 1], 2, 1, 3) == F21 / 3
     F21_new = (1 * 0.5 * 2^2) / (1 + 0.5 * 1 * 2^2)
-    @test Fclassic1_fw([2, 1], 2, 1, 1) == F21_new # ...but depend on res. mass
+    @test Fclassic1_fw([2.0, 1], 2, 1, 1) == F21_new # ...but depend on res. mass
 
     # Index by index - MultiplexNetwork
     Fclassic1_nti = ClassicResponse(multi_net1; hₜ = 1.0, aᵣ = 0.5)
-    @test Fclassic1_nti([1, 1], 1, 1, 1) == 0 # no interaction
-    @test Fclassic1_nti([1, 1], 1, 2, 1) == 0 # no interaction
-    @test Fclassic1_nti([1, 1], 2, 2, 1) == 0 # no interaction
+    @test Fclassic1_nti([1.0, 1], 1, 1, 1) == 0 # no interaction
+    @test Fclassic1_nti([1.0, 1], 1, 2, 1) == 0 # no interaction
+    @test Fclassic1_nti([1.0, 1], 2, 2, 1) == 0 # no interaction
     F21 = (1 * 0.5 * 1^2) / (1 + 0.5 * 1 * 1^2)
-    @test Fclassic1_nti([1, 1], 2, 1, 1) == F21 # interaction
-    @test Fclassic1_nti([1, 2], 2, 1, 1) == F21 # don't depend on cons. mass
-    @test Fclassic1_nti([1, 1], 2, 1, 2) == F21 / 2
-    @test Fclassic1_nti([1, 1], 2, 1, 3) == F21 / 3
+    @test Fclassic1_nti([1.0, 1], 2, 1, 1) == F21 # interaction
+    @test Fclassic1_nti([1.0, 2], 2, 1, 1) == F21 # don't depend on cons. mass
+    @test Fclassic1_nti([1.0, 1], 2, 1, 2) == F21 / 2
+    @test Fclassic1_nti([1.0, 1], 2, 1, 3) == F21 / 3
     F21_new = (1 * 0.5 * 2^2) / (1 + 0.5 * 1 * 2^2)
-    @test Fclassic1_nti([2, 1], 2, 1, 1) == F21_new # ...but depend on res. mass
+    @test Fclassic1_nti([2.0, 1], 2, 1, 1) == F21_new # ...but depend on res. mass
 
     # Matrix
-    @test Fclassic1_fw([1, 1], foodweb_2sp) == sparse([0 0; F21 0]) # provide biomass vector
-    @test Fclassic1_fw(1, foodweb_2sp) == sparse([0 0; F21 0]) # or a scalar
-    @test Fclassic1_nti([1, 1], multi_net1) == sparse([0 0; F21 0]) # provide biomass vector
-    @test Fclassic1_nti(1, multi_net1) == sparse([0 0; F21 0]) # or a scalar
+    @test Fclassic1_fw([1.0, 1], foodweb_2sp) == sparse([0 0; F21 0]) # provide biomass vector
+    @test Fclassic1_fw(1.0, foodweb_2sp) == sparse([0 0; F21 0]) # or a scalar
+    @test Fclassic1_nti([1.0, 1], multi_net1) == sparse([0 0; F21 0]) # provide biomass vector
+    @test Fclassic1_nti(1.0, multi_net1) == sparse([0 0; F21 0]) # or a scalar
 
     # Non-default hill exponent
     Fclassic_1 = ClassicResponse(foodweb_2sp; h = 3, hₜ = 1.0, aᵣ = 0.5)
     F21 = (1 * 0.5 * 2^3) / (1 + 0.5 * 1 * 2^3)
-    @test Fclassic_1(2, foodweb_2sp) == sparse([0 0; F21 0])
+    @test Fclassic_1(2.0, foodweb_2sp) == sparse([0 0; F21 0])
 
     # Non-default body mass
     foodweb_2sp_M = FoodWeb(A_2sp; M = [2, 3])
     Fclassic_1 = ClassicResponse(foodweb_2sp; h = 3, hₜ = 1.0, aᵣ = 0.5)
     F21 = (1 * 0.5 * 2^3) / (1 + 0.5 * 1 * 2^3)
-    @test Fclassic_1(2, foodweb_2sp_M) == sparse([0 0; F21/3 0])
+    @test Fclassic_1(2.0, foodweb_2sp_M) == sparse([0 0; F21/3 0])
 
     # Non-default attack rate
     Fclassic_1 = ClassicResponse(foodweb_2sp; aᵣ = 0.2, hₜ = 1.0)
     F21 = (1 * 0.2 * 2^2) / (1 + 0.2 * 1 * 2^2)
-    @test Fclassic_1(2, foodweb_2sp) == sparse([0 0; F21 0])
+    @test Fclassic_1(2.0, foodweb_2sp) == sparse([0 0; F21 0])
 
     # Non-default handling time
     Fclassic_1 = ClassicResponse(foodweb_2sp; hₜ = 2, aᵣ = 0.5)
     F21 = (1 * 0.5 * 2^2) / (1 + 0.5 * 2 * 2^2)
-    @test Fclassic_1(2, foodweb_2sp) == sparse([0 0; F21 0])
+    @test Fclassic_1(2.0, foodweb_2sp) == sparse([0 0; F21 0])
 
 
     # Consumer feeding on several resources
     Fclassic2_fw = ClassicResponse(foodweb_3sp; hₜ = 1.0, aᵣ = 0.5)
     Fclassic2_nti = ClassicResponse(net_interference; hₜ = 1.0, aᵣ = 0.5)
-    B = [1, 1, 1] # uniform biomass distribution
+    B = [1.0, 1, 1] # uniform biomass distribution
     F21 = (1 * 0.5 * 1^2) / (1 + 0.5 * 1 * 1^2)
     F31 = (0.5 * 0.5 * 1^2) / (1 + 0.5 * 0.5 * 1 * 1^2 + 0.5 * 0.5 * 1 * 1^2)
     F32 = F31
     @test Fclassic2_fw(B, foodweb_3sp) == sparse([0 0 0; F21 0 0; F31 F32 0])
     @test Fclassic2_nti(B, net_interference) == sparse([0 0 0; F21 0 0; F31 F32 0])
-    B = [3, 2, 1] # non-uniform biomass distribution
+    B = [3.0, 2, 1] # non-uniform biomass distribution
     F21 = (1 * 0.5 * 3^2) / (1 + 0.5 * 1 * 3^2)
     F31 = (0.5 * 0.5 * 3^2) / (1 + 0.5 * 0.5 * 1 * 3^2 + 0.5 * 0.5 * 1 * 2^2)
     F32 = (0.5 * 0.5 * 2^2) / (1 + 0.5 * 0.5 * 1 * 3^2 + 0.5 * 0.5 * 1 * 2^2)
     @test Fclassic2_fw(B, foodweb_3sp) == sparse([0 0 0; F21 0 0; F31 F32 0])
     @test Fclassic2_nti(B, net_interference) == sparse([0 0 0; F21 0 0; F31 F32 0])
-    B, aᵣ = [3, 2, 1], [0 0 0; 0.5 0 0; 0.5 0.2 0] # non-uniform biomass...
+    B, aᵣ = [3.0, 2, 1], [0 0 0; 0.5 0 0; 0.5 0.2 0] # non-uniform biomass...
     Fclassic2_fw = ClassicResponse(foodweb_3sp; aᵣ = aᵣ, hₜ = 1.0) #...and non-uniform attack rate
     Fclassic2_nti = ClassicResponse(net_interference; aᵣ = aᵣ, hₜ = 1.0)
     F21 = (1 * 0.5 * 3^2) / (1 + 0.5 * 1 * 3^2)
@@ -262,7 +262,7 @@ end
     F32 = (0.5 * 0.2 * 2^2) / (1 + 0.5 * 0.5 * 1 * 3^2 + 0.5 * 0.2 * 1 * 2^2)
     @test Fclassic2_fw(B, foodweb_3sp) == sparse([0 0 0; F21 0 0; F31 F32 0])
     @test Fclassic2_nti(B, net_interference) == sparse([0 0 0; F21 0 0; F31 F32 0])
-    B, hₜ = [3, 2, 1], [0 0 0; 0.9 0 0; 0.7 0.2 0] # non-uniform biomass...
+    B, hₜ = [3.0, 2, 1], [0 0 0; 0.9 0 0; 0.7 0.2 0] # non-uniform biomass...
     Fclassic2_fw = ClassicResponse(foodweb_3sp; hₜ = hₜ, aᵣ = 0.5) #...+ varying handling time
     Fclassic2_nti = ClassicResponse(net_interference; hₜ = hₜ, aᵣ = 0.5)
     F21 = (1 * 0.5 * 3^2) / (1 + 0.5 * 0.9 * 3^2)
@@ -275,7 +275,7 @@ end
     # Adding intraspecific interference
     Fclassic2_fw = ClassicResponse(foodweb_3sp; c = 1, hₜ = 1.0, aᵣ = 0.5)
     Fclassic2_nti = ClassicResponse(net_interference; c = 1, hₜ = 1.0, aᵣ = 0.5)
-    B = [3, 2, 1] # non-uniform biomass distribution
+    B = [3.0, 2, 1] # non-uniform biomass distribution
     F21 = (1 * 0.5 * 3^2) / (1 + 1 * 2 + 0.5 * 1 * 3^2)
     F31 = (0.5 * 0.5 * 3^2) / (1 + 1 * 1 + 0.5 * 0.5 * 1 * 3^2 + 0.5 * 0.5 * 1 * 2^2)
     F32 = (0.5 * 0.5 * 2^2) / (1 + 1 * 1 + 0.5 * 0.5 * 1 * 3^2 + 0.5 * 0.5 * 1 * 2^2)
@@ -285,7 +285,7 @@ end
     # Adding interspecific interference
     net_interference.layers[:interference].intensity = 0.6 # activate interspecific interference
     Fclassic2_nti = ClassicResponse(net_interference; c = 0.5, hₜ = 1.0, aᵣ = 0.5)
-    B = [3, 2, 1] # non-uniform biomass distribution
+    B = [3.0, 2, 1] # non-uniform biomass distribution
     F21 = (1 * 0.5 * 3^2) / (1 + 0.5 * 2 + 0.6 * 1 + 0.5 * 1 * 3^2)
     F31 =
         (0.5 * 0.5 * 3^2) /
@@ -298,7 +298,7 @@ end
     # Adding refuge provisioning
     Fclassic2_nti = ClassicResponse(net_refuge; c = 0.0, hₜ = 1.0, aᵣ = 0.5)
     Fclassic2_fw = ClassicResponse(foodweb_3sp; c = 0.0, hₜ = 1.0, aᵣ = 0.5)
-    B = [3, 2, 1]
+    B = [3.0, 2, 1]
     @test Fclassic2_fw(B, foodweb_3sp) == Fclassic2_nti(B, net_refuge) # nti intensity = 0 <=> food web
     for r0 in [0.1, 0.2, 0.25]
         net_refuge.layers[:refuge].intensity = r0